requirements.txt
# Framework principal
streamlit==1.39.0
pandas>=2.1.0,<2.3.0
numpy>=1.26.0,<2.0.0
Pillow>=10.0.0

# OpenAI
openai>=1.57.0

# Visualización
plotly>=5.18.0
openpyxl>=3.1.0

# ML básico
scikit-learn>=1.4.0
scipy>=1.13.0

# Evaluación de imágenes (versiones compatibles con Streamlit Cloud)
torch==2.1.2
torchvision==0.16.2
lpips==0.1.4
pytorch-msssim==1.0.0
open-clip-torch==2.24.0

# Utilidades
python-dotenv>=1.0.0